git-fixup: Learn -i interactive mode
Almost always the target to fixup can be found by the
git-fixup-targets. Having to select the target by copying the hash
and then pasting it into a git-fixup command is more tedious than it
really has to be.

This change ensures that the git-fixup script understand the -i flag
which will list all the fixup-targets with associated numbers, it then
output a prompt for the user. If the user enters the corresponding
number a fixup commit for the target will be created.